REL-1177: moved ItcOverheadProvider
